+++ wazo_confd/plugins/meeting/plugin.py
@@ -16,6 +16,9 @@ from wazo_confd.plugins.endpoint_sip.service import (
 from wazo_confd.plugins.endpoint_sip.service import (
     build_template_service as build_endpoint_sip_template_service,
 )
+from wazo_confd.plugins.extension_feature.service import (
+    build_service as build_extension_features_service,
+)
 from wazo_confd.plugins.ingress_http.service import (
     build_service as build_ingress_http_service,
 )
@@ -36,14 +39,15 @@ class Plugin:
         pjsip_doc = dependencies['pjsip_doc']
 
         ingress_http_service = build_ingress_http_service()
-        api_notifier = Notifier(bus, sysconfd, ingress_http_service)
+        extension_features_service = build_extension_features_service()
+        api_notifier = Notifier(bus, sysconfd, ingress_http_service, extension_features_service)
         service = build_service(api_notifier)
 
         instant_bus_publisher = InstantBusPublisher.from_config(
             config['bus'],
             config['uuid'],
         )
-        bus_notifier = Notifier(instant_bus_publisher, sysconfd, ingress_http_service)
+        bus_notifier = Notifier(instant_bus_publisher, sysconfd, ingress_http_service, extension_features_service)
         bus_event_handler = MeetingBusEventHandler(service, bus_notifier)
         bus_event_handler.subscribe(bus_consumer)
 
@@ -53,7 +57,7 @@ class Plugin:
         )
         user_service = build_user_service(provd_client=None)
         tenant_service = build_tenant_service()
-        args = [service, user_service, ingress_http_service]
+        args = [service, user_service, ingress_http_service, extension_features_service]
 
         api.add_resource(
             MeetingList,
@@ -65,6 +69,7 @@ class Plugin:
                 endpoint_sip_service,
                 endpoint_sip_template_service,
                 ingress_http_service,
+                extension_features_service,
             ],
         )
         api.add_resource(
@@ -95,6 +100,7 @@ class Plugin:
                 endpoint_sip_service,
                 endpoint_sip_template_service,
                 ingress_http_service,
+                extension_features_service,
                 auth_client,
             ],
         )